Add adobe-edge to e2e app
