+++ apps/e2e/.env.example
@@ -10,3 +10,13 @@ CONVIVA_VIEWER_ID=e2e_viewer_id
 BITMOVIN_LICENSE_KEY=
 BITMOVIN_LOG_LEVEL=DEBUG
 BITMOVIN_CUSTOM_USER_ID=connector_e2e_user
+
+# YOSPACE
+YOSPACE_USERNAME=
+YOSPACE_PASSWORD=
+
+# ADOBE-EDGE
+ADOBE_EDGE_WEB_DATASTREAM_ID=fill-in-your-datastream-id
+ADOBE_EDGE_WEB_BASE_PATH=ee-pre-prd
+ADOBE_EDGE_WEB_ORG_ID=ADB3LETTERSANDNUMBERS@AdobeOrg
+ADOBE_EDGE_MOBILE_ENV_ID=fill-in-your-env-id
